restructure the files and classes
* move the files into a `lib` directory
* split the classes out of the main rogue.py file into one class per
  file under the lib/rogue dir
* put the constants into lib/rogue/constants
